+++ frontend/src/features/archived-items/archived-item-list/index.ts
@@ -1,0 +1,2 @@
+import "./archived-item-list";
+import "./archived-item-list-item";
