+++ frontend/src/features/archived-items/archived-item-list/types.ts
@@ -1,0 +1,3 @@
+export type CheckboxChangeEventDetail = {
+  checked: boolean;
+};
